Simplified collector names, so that they do not hide the general nature of those classes.
