+++ graylog-storage-opensearch2/src/main/java/org/graylog/storage/opensearch2/views/OpenSearchBackend.java
@@ -28,7 +28,7 @@ import org.graylog.plugins.views.search.elasticsearch.IndexLookup;
 import org.graylog.plugins.views.search.engine.BackendQuery;
 import org.graylog.plugins.views.search.engine.QueryBackend;
 import org.graylog.plugins.views.search.engine.QueryExecutionStats;
-import org.graylog.plugins.views.search.engine.monitoring.collection.QueryExecutionStatsCollector;
+import org.graylog.plugins.views.search.engine.monitoring.collection.StatsCollector;
 import org.graylog.plugins.views.search.errors.SearchError;
 import org.graylog.plugins.views.search.errors.SearchTypeError;
 import org.graylog.plugins.views.search.errors.SearchTypeErrorParser;
@@ -79,7 +79,7 @@ public class OpenSearchBackend implements QueryBackend<OSGeneratedQueryContext> 
     private final OSGeneratedQueryContext.Factory queryContextFactory;
     private final UsedSearchFiltersToQueryStringsMapper usedSearchFiltersToQueryStringsMapper;
     private final boolean allowLeadingWildcard;
-    private final QueryExecutionStatsCollector<QueryExecutionStats> executionStatsCollector;
+    private final StatsCollector<QueryExecutionStats> executionStatsCollector;
 
     @Inject
     public OpenSearchBackend(Map<String, Provider<OSSearchTypeHandler<? extends SearchType>>> elasticsearchSearchTypeHandlers,
@@ -87,7 +87,7 @@ public class OpenSearchBackend implements QueryBackend<OSGeneratedQueryContext> 
                              IndexLookup indexLookup,
                              OSGeneratedQueryContext.Factory queryContextFactory,
                              UsedSearchFiltersToQueryStringsMapper usedSearchFiltersToQueryStringsMapper,
-                             QueryExecutionStatsCollector<QueryExecutionStats> executionStatsCollector,
+                             StatsCollector<QueryExecutionStats> executionStatsCollector,
                              @Named("allow_leading_wildcard_searches") boolean allowLeadingWildcard) {
         this.openSearchSearchTypeHandlers = elasticsearchSearchTypeHandlers;
         this.client = client;
@@ -106,7 +106,7 @@ public class OpenSearchBackend implements QueryBackend<OSGeneratedQueryContext> 
     }
 
     @Override
-    public QueryExecutionStatsCollector<QueryExecutionStats> getExecutionStatsCollector() {
+    public StatsCollector<QueryExecutionStats> getExecutionStatsCollector() {
         return this.executionStatsCollector;
     }
