+++ graylog2-server/src/main/java/org/graylog2/storage/VersionAwareStorageModule.java
@@ -23,9 +23,9 @@ import org.graylog.plugins.views.migrations.V20200730000000_AddGl2MessageIdField
 import org.graylog.plugins.views.search.engine.GeneratedQueryContext;
 import org.graylog.plugins.views.search.engine.QueryBackend;
 import org.graylog.plugins.views.search.engine.QueryExecutionStats;
-import org.graylog.plugins.views.search.engine.monitoring.collection.InMemoryCappedQueryExecutionStatsCollector;
-import org.graylog.plugins.views.search.engine.monitoring.collection.NoOpExecutionStatsCollector;
-import org.graylog.plugins.views.search.engine.monitoring.collection.QueryExecutionStatsCollector;
+import org.graylog.plugins.views.search.engine.monitoring.collection.InMemoryCappedStatsCollector;
+import org.graylog.plugins.views.search.engine.monitoring.collection.NoOpStatsCollector;
+import org.graylog.plugins.views.search.engine.monitoring.collection.StatsCollector;
 import org.graylog2.Configuration;
 import org.graylog2.indexer.IndexToolsAdapter;
 import org.graylog2.indexer.cluster.ClusterAdapter;
@@ -88,11 +88,11 @@ public class VersionAwareStorageModule extends AbstractModule {
 
     private void bindQueryBackend() {
         if (configuration.isQueryLatencyMonitoringEnabled() && configuration.getQueryLatencyMonitoringWindowSize() > 0) {
-            bind(new TypeLiteral<QueryExecutionStatsCollector<QueryExecutionStats>>() {})
-                    .toInstance(new InMemoryCappedQueryExecutionStatsCollector<>(configuration.getQueryLatencyMonitoringWindowSize()));
+            bind(new TypeLiteral<StatsCollector<QueryExecutionStats>>() {})
+                    .toInstance(new InMemoryCappedStatsCollector<>(configuration.getQueryLatencyMonitoringWindowSize()));
         } else {
-            bind(new TypeLiteral<QueryExecutionStatsCollector<QueryExecutionStats>>() {})
-                    .toInstance(new NoOpExecutionStatsCollector<>());
+            bind(new TypeLiteral<StatsCollector<QueryExecutionStats>>() {})
+                    .toInstance(new NoOpStatsCollector<>());
         }
         bind(new TypeLiteral<QueryBackend<? extends GeneratedQueryContext>>() {})
                 .toProvider(ElasticsearchBackendProvider.class);
